requirements.txt
Flask
gunicorn
tensorflow-cpu==2.13.0
numpy
Pillow
gdown
